Fixed one-way street display and annotation mismatch.
The plotly.graph_objs barmode attribute was modified to only be 'group' if there were two directions. The annotations are added based on presence of Eastbound or Westbound data. The graphs now look much better but could still use some basic formatting, and the annotations are displaying the wrong text.
Issue #2