Analysed suspicious baselines.
For each outlier a scatter plot of the week, the baselines overlaid on percetile bands, and the baselines with the outliers removed were plotted.
Each set of figures has a brief summarising comment detailing the day removed, the approximate time the removal of the date affected and to what degree, and whether the day was on a weekend or weekday.
Finally, the instances where removing the outliers had a significant enough impact on the baseline were left open, and the rest of the windows were closed.