Tweaked display.
Now with descriptions of the segments before the graphs. If the date has been removed it is stated explicitly at the end of the description, and there is a list of dates that have been removed on any given segment at the end of the notebook, and in the text file.